requirements.txt
Flask
langchain
langchain_community
python-dotenv
Chroma
sentence-transformers
